microservices/measurement-based-care/pyproject.toml
[tool.poetry]
name = "measurement-based-care"
version = "0.1.0"
description = ""
authors = ["Vijay Selvaraj <vijay.selvaraj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
transitions = "^0.9.0"
pyyaml = "^6.0.1"
fastapi = "^0.110.2"
mangum = "^0.17.0"
aws-lambda-powertools = "^2.37.0"
uvicorn = "^0.29.0"
jmespath = "^1.0.1"
pydantic = {extras = ["email"], version = "^2.7.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
